Projetos Maiores/Qr Code API - Python (Flask)/requirements.txt

pyqrcode
flask
flask_cors
pyzbar
opencv-python